Add gradebook uradjen i jos par stvari

+++ routes/api.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\AuthController;
 use App\Http\Controllers\CommentController;
 use App\Http\Controllers\GradebookController;
 use App\Http\Controllers\UserController;
@@ -17,26 +18,32 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get('/users', [UserController::class, 'index']);
-Route::get('/users/{id}', [UserController::class, 'show']);
-Route::post('/users', [UserController::class, 'store']);
-Route::delete('/users/{id}', [UserController::class, 'destroy']);
-Route::put('/users/{id}', [UserController::class, 'update']);
+Route::post('/login', [AuthController::class, 'login']);
+Route::post('/register', [AuthController::class, 'register']);
+Route::get('/active-user', [AuthController::class, 'getActiveUser'])->middleware('auth');
+Route::post('/logout', [AuthController::class, 'logout'])->middleware('auth');
 
-Route::get('/gradebooks', [GradebookController::class, 'index']);
-Route::get('/gradebooks/{id}', [GradebookController::class, 'show']);
-Route::post('/gradebooks', [GradebookController::class, 'store']);
-Route::delete('/gradebooks/{id}', [GradebookController::class, 'destroy']);
-Route::put('/gradebooks/{id}', [GradebookController::class, 'update']);
+Route::get('/users', [UserController::class, 'index'])->middleware('auth');
+Route::get('/users/available', [UserController::class, 'getAvailable'])->middleware(('auth'));
+Route::get('/users/{id}', [UserController::class, 'show'])->middleware('auth');
+Route::post('/users', [UserController::class, 'store'])->middleware('auth');
+Route::delete('/users/{id}', [UserController::class, 'destroy'])->middleware('auth');
+Route::put('/users/{id}', [UserController::class, 'update'])->middleware('auth');
 
-Route::get('/comments', [CommentController::class, 'index']);
-Route::get('/comments/{id}', [CommentController::class, 'show']);
-Route::post('/comments', [CommentController::class, 'store']);
-Route::delete('/comments/{id}', [CommentController::class, 'destroy']);
-Route::put('/comments/{id}', [CommentController::class, 'update']);
+Route::get('/gradebooks', [GradebookController::class, 'index'])->middleware('auth');
+Route::get('/gradebooks/{id}', [GradebookController::class, 'show'])->middleware('auth');
+Route::post('/gradebooks', [GradebookController::class, 'store'])->middleware('auth');
+Route::delete('/gradebooks/{id}', [GradebookController::class, 'destroy'])->middleware('auth');
+Route::put('/gradebooks/{id}', [GradebookController::class, 'update'])->middleware('auth');
 
-Route::get('/students', [StudentController::class, 'index']);
-Route::get('/students/{id}', [StudentController::class, 'show']);
-Route::post('/students', [StudentController::class, 'store']);
-Route::delete('/students/{id}', [StudentController::class, 'destroy']);
-Route::put('/students/{id}', [StudentController::class, 'update']);
+Route::get('/comments', [CommentController::class, 'index'])->middleware('auth');
+Route::get('/comments/{id}', [CommentController::class, 'show'])->middleware('auth');
+Route::post('/comments', [CommentController::class, 'store'])->middleware('auth');
+Route::delete('/comments/{id}', [CommentController::class, 'destroy'])->middleware('auth');
+Route::put('/comments/{id}', [CommentController::class, 'update'])->middleware('auth');
+
+Route::get('/students', [StudentController::class, 'index'])->middleware('auth');
+Route::get('/students/{id}', [StudentController::class, 'show'])->middleware('auth');
+Route::post('/students', [StudentController::class, 'store'])->middleware('auth');
+Route::delete('/students/{id}', [StudentController::class, 'destroy'])->middleware('auth');
+Route::put('/students/{id}', [StudentController::class, 'update'])->middleware('auth');
